+++ Parylation/Parylation/Presentation/Scenes/Home/HomeContract.swift
@@ -13,6 +13,8 @@ import UIKit
 
 protocol HomeContainer {
     var taskRepository: TaskRepository { get }
+    var iconRepository: IconRepository { get }
+    var colorRepository: ColorRepository { get }
 }
 
 protocol HomeBuilder: AnyObject {
@@ -27,8 +29,11 @@ protocol HomeRouter: AnyObject {
 protocol HomeViewModel {
     var reloadTrigger: AnyObserver<Void> { get }
     var createTrigger: AnyObserver<Void> { get }
+    var willDisplayItemTrigger: AnyObserver<IndexPath> { get }
     var selectTrigger: AnyObserver<IndexPath> { get }
     var deleteTrigger: AnyObserver<IndexPath> { get }
-
+    
+    var itemIcon: Driver<(Icon, IndexPath)> { get }
+    var itemColor: Driver<(ParylationDomain.Color, IndexPath)> { get }
     var sections: Driver<[HomeTableSection]> { get }
 }