+++ ParylationDomain/ParylationDomain/Interactors/HomeInteractor.swift
@@ -14,6 +14,8 @@ public enum HomeInteractorError: Error {
 
 public protocol HomeInteractor {
     func fetchTaks() -> Single<[Task]>
+    func fetchIcon(id: String) -> Single<Icon>
+    func fetchColor(id: String) -> Single<Color>
     func deleteTask(taskId: String) -> Single<Void>
 
     // TEMP:
@@ -22,9 +24,17 @@ public protocol HomeInteractor {
 
 public final class HomeInteractorImpl {
     private let taskRepository: TaskRepository
+    private let iconRepository: IconRepository
+    private let colorRepository: ColorRepository
 
-    public init(taskRepository: TaskRepository) {
+    public init(
+        taskRepository: TaskRepository,
+        iconRepository: IconRepository,
+        colorRepository: ColorRepository
+    ) {
         self.taskRepository = taskRepository
+        self.iconRepository = iconRepository
+        self.colorRepository = colorRepository
     }
 }
 
@@ -33,6 +43,16 @@ public final class HomeInteractorImpl {
 extension HomeInteractorImpl: HomeInteractor {
     public func fetchTaks() -> Single<[Task]> {
         return taskRepository.fetchTasks()
+            .catchError { _ in .error(HomeInteractorError.failed) }
+    }
+
+    public func fetchIcon(id: String) -> Single<Icon> {
+        return iconRepository.fetch(id: id)
+            .catchError { _ in .error(HomeInteractorError.failed) }
+    }
+
+    public func fetchColor(id: String) -> Single<Color> {
+        return colorRepository.fetch(id: id)
             .catchError { _ in .error(HomeInteractorError.failed) }
     }